HDDS-5788. Reduce run time for TestOzoneManagerHA tests
