+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestOzoneManagerHA.java
@@ -37,8 +37,10 @@ import org.apache.hadoop.ozone.client.rpc.RpcClient;
 import org.apache.hadoop.ozone.om.ha.OMFailoverProxyProvider;
 import org.apache.hadoop.ozone.om.ratis.OzoneManagerRatisServerConfig;
 import org.apache.ozone.test.GenericTestUtils;
+import org.junit.BeforeClass;
 import org.junit.Before;
 import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Rule;
 import org.junit.Assert;
 
@@ -50,6 +52,7 @@ import java.net.ConnectException;
 import java.time.Duration;
 import java.util.UUID;
 import java.util.HashMap;
+import java.util.concurrent.TimeoutException;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.apache.hadoop.fs.CommonConfigurationKeysPublic.IPC_CLIENT_CONNECT_MAX_RETRIES_KEY;
@@ -68,14 +71,14 @@ import static org.junit.Assert.fail;
  */
 public abstract class TestOzoneManagerHA {
 
-  private MiniOzoneOMHAClusterImpl cluster = null;
-  private MiniOzoneCluster.Builder clusterBuilder = null;
-  private ObjectStore objectStore;
-  private OzoneConfiguration conf;
-  private String clusterId;
-  private String scmId;
-  private String omId;
-  private String omServiceId;
+  private static MiniOzoneOMHAClusterImpl cluster = null;
+  private static MiniOzoneCluster.Builder clusterBuilder = null;
+  private static ObjectStore objectStore;
+  private static OzoneConfiguration conf;
+  private static String clusterId;
+  private static String scmId;
+  private static String omId;
+  private static String omServiceId;
   private static int numOfOMs = 3;
   private static final int LOG_PURGE_GAP = 50;
   /* Reduce max number of retries to speed up unit test. */
@@ -88,7 +91,7 @@ public abstract class TestOzoneManagerHA {
   public ExpectedException exception = ExpectedException.none();
 
   @Rule
-  public Timeout timeout = Timeout.seconds(300);;
+  public Timeout timeout = Timeout.seconds(300);
 
   public MiniOzoneOMHAClusterImpl getCluster() {
     return cluster;
@@ -137,8 +140,8 @@ public abstract class TestOzoneManagerHA {
    *
    * @throws IOException
    */
-  @Before
-  public void init() throws Exception {
+  @BeforeClass
+  public static void init() throws Exception {
     conf = new OzoneConfiguration();
     clusterId = UUID.randomUUID().toString();
     scmId = UUID.randomUUID().toString();
@@ -171,7 +174,6 @@ public abstract class TestOzoneManagerHA {
      */
     conf.set(OZONE_BLOCK_DELETING_SERVICE_INTERVAL, "10s");
     conf.set(OZONE_KEY_DELETING_LIMIT_PER_TASK, "2");
-    additionalConfiguration();
 
     clusterBuilder = MiniOzoneCluster.newOMHABuilder(conf)
         .setClusterId(clusterId)
@@ -179,12 +181,20 @@ public abstract class TestOzoneManagerHA {
         .setOMServiceId(omServiceId)
         .setOmId(omId)
         .setNumOfOzoneManagers(numOfOMs);
-    additionalClusterSettings();
 
-    cluster = (MiniOzoneOMHAClusterImpl)clusterBuilder.build();
+    cluster = (MiniOzoneOMHAClusterImpl) clusterBuilder.build();
     cluster.waitForClusterToBeReady();
     objectStore = OzoneClientFactory.getRpcClient(omServiceId, conf)
         .getObjectStore();
+  }
+
+  /**
+   * Apply additional configuration between tests.
+   */
+  @Before
+  public void setupTest() {
+    additionalConfiguration();
+    additionalClusterSettings();
   }
 
   /**
@@ -202,10 +212,21 @@ public abstract class TestOzoneManagerHA {
   }
 
   /**
-   * Shutdown MiniDFSCluster.
+   * Reset cluster between tests.
    */
   @After
-  public void shutdown() {
+  public void resetCluster()
+      throws IOException {
+    if (cluster != null) {
+      cluster.restartOzoneManager();
+    }
+  }
+
+  /**
+   * Shutdown MiniDFSCluster after all tests of a class have run.
+   */
+  @AfterClass
+  public static void shutdown() {
     if (cluster != null) {
       cluster.shutdown();
     }
@@ -213,6 +234,7 @@ public abstract class TestOzoneManagerHA {
 
   /**
    * Create a key in the bucket.
+   *
    * @return the key name.
    */
   public static String createKey(OzoneBucket ozoneBucket) throws IOException {
@@ -256,6 +278,7 @@ public abstract class TestOzoneManagerHA {
 
   /**
    * Stop the current leader OM.
+   *
    * @throws Exception
    */
   protected void stopLeaderOM() {
@@ -316,6 +339,7 @@ public abstract class TestOzoneManagerHA {
   /**
    * This method createFile and verifies the file is successfully created or
    * not.
+   *
    * @param ozoneBucket
    * @param keyName
    * @param data
@@ -324,7 +348,8 @@ public abstract class TestOzoneManagerHA {
    * @throws Exception
    */
   protected void testCreateFile(OzoneBucket ozoneBucket, String keyName,
-      String data, boolean recursive, boolean overwrite)
+                                String data, boolean recursive,
+                                boolean overwrite)
       throws Exception {
 
     OzoneOutputStream ozoneOutputStream = ozoneBucket.createFile(keyName,
@@ -347,6 +372,62 @@ public abstract class TestOzoneManagerHA {
     byte[] fileContent = new byte[data.getBytes(UTF_8).length];
     ozoneInputStream.read(fileContent);
     Assert.assertEquals(data, new String(fileContent, UTF_8));
+  }
+
+  protected void createKeyTest(boolean checkSuccess) throws Exception {
+    String userName = "user" + RandomStringUtils.randomNumeric(5);
+    String adminName = "admin" + RandomStringUtils.randomNumeric(5);
+    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
+
+    VolumeArgs createVolumeArgs = VolumeArgs.newBuilder()
+        .setOwner(userName)
+        .setAdmin(adminName)
+        .build();
+
+    try {
+      getObjectStore().createVolume(volumeName, createVolumeArgs);
+
+      OzoneVolume retVolumeinfo = getObjectStore().getVolume(volumeName);
+
+      Assert.assertTrue(retVolumeinfo.getName().equals(volumeName));
+      Assert.assertTrue(retVolumeinfo.getOwner().equals(userName));
+      Assert.assertTrue(retVolumeinfo.getAdmin().equals(adminName));
+
+      String bucketName = UUID.randomUUID().toString();
+      String keyName = UUID.randomUUID().toString();
+      retVolumeinfo.createBucket(bucketName);
+
+      OzoneBucket ozoneBucket = retVolumeinfo.getBucket(bucketName);
+
+      Assert.assertTrue(ozoneBucket.getName().equals(bucketName));
+      Assert.assertTrue(ozoneBucket.getVolumeName().equals(volumeName));
+
+      String value = "random data";
+      OzoneOutputStream ozoneOutputStream = ozoneBucket.createKey(keyName,
+          value.length(), ReplicationType.STAND_ALONE,
+          ReplicationFactor.ONE, new HashMap<>());
+      ozoneOutputStream.write(value.getBytes(UTF_8), 0, value.length());
+      ozoneOutputStream.close();
+
+      OzoneInputStream ozoneInputStream = ozoneBucket.readKey(keyName);
+
+      byte[] fileContent = new byte[value.getBytes(UTF_8).length];
+      ozoneInputStream.read(fileContent);
+      Assert.assertEquals(value, new String(fileContent, UTF_8));
+
+    } catch (ConnectException | RemoteException e) {
+      if (!checkSuccess) {
+        // If the last OM to be tried by the RetryProxy is down, we would get
+        // ConnectException. Otherwise, we would get a RemoteException from the
+        // last running OM as it would fail to get a quorum.
+        if (e instanceof RemoteException) {
+          GenericTestUtils.assertExceptionContains(
+              "OMNotLeaderException", e);
+        }
+      } else {
+        throw e;
+      }
+    }
   }
 
 }
